+++ packages/backend/src/engine/nested_stack_resolver.ts
@@ -1,10 +1,13 @@
 import { NestedStack, Stack } from 'aws-cdk-lib';
+import { AttributionMetadataStorage } from '@aws-amplify/backend-output-storage';
+import { fileURLToPath } from 'url';
 
 /**
  * Vends stacks for a resource grouping
  */
 export type StackResolver = {
   getStackFor: (resourceGroupName: string) => Stack;
+  getCustomStack: (name: string) => Stack;
 };
 
 /**
@@ -16,7 +19,24 @@ export class NestedStackResolver implements StackResolver {
   /**
    * Initialize with a root stack
    */
-  constructor(private readonly rootStack: Stack) {}
+  constructor(
+    private readonly rootStack: Stack,
+    private readonly attributionMetadataStorage: AttributionMetadataStorage
+  ) {}
+
+  /**
+   * Proxy to getStackFor that appends attribution metadata for custom stacks
+   */
+  getCustomStack = (name: string): Stack => {
+    const stack = this.getStackFor(name);
+    // this is safe even if stack is cached from an earlier invocation because storeAttributionMetadata is a noop if the stack description already exists
+    this.attributionMetadataStorage.storeAttributionMetadata(
+      stack,
+      `custom`,
+      fileURLToPath(new URL('../../package.json', import.meta.url))
+    );
+    return stack;
+  };
 
   /**
    * Returns a cached NestedStack if resourceGroupName has been seen before