+++ packages/plugin-types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @aws-amplify/plugin-types
 
+## 0.3.1
+
+### Patch Changes
+
+- 3bff764b: Expose user pool and client as cfn resources.
+
 ## 0.3.0
 
 ### Minor Changes